Place duration in a seperate file
